Moved the update offset method out of the filesystemseedvaluereader class and into it's own class.
